+++ features/step_definitions/portlets_steps.rb
@@ -21,7 +21,7 @@ end
 
 Given /^a page with a portlet that raises a Not Found exception exists$/ do
   @last_page = create(:public_page)
-  @raises_not_found = create(:portlet, :code => 'raise ActiveRecord::RecordNotFound', :template=>"I shouldn't be shown.'")
+  @raises_not_found = create(:portlet, :code => 'raise ActiveRecord::RecordNotFound', :template => "I shouldn't be shown.'")
   @last_page.add_content(@raises_not_found)
   @last_page.publish!
   assert @last_page.published?
@@ -29,7 +29,7 @@ end
 
 When /^a page with a portlet that raises an Access Denied exception exists$/ do
   @last_page = create(:public_page)
-  @raises_access_denied = create(:portlet, :code => 'raise Cms::Errors::AccessDenied', :template=>"I shouldn't be shown.'")
+  @raises_access_denied = create(:portlet, :code => 'raise Cms::Errors::AccessDenied', :template => "I shouldn't be shown.'")
   @last_page.add_content(@raises_access_denied)
   @last_page.publish!
   assert @last_page.published?
@@ -37,7 +37,7 @@ end
 
 When /^a page with a portlet that display "([^"]*)" exists$/ do |view|
   @last_page = create(:public_page)
-  portlet = create(:portlet, :template=>view)
+  portlet = create(:portlet, :template => view)
   @last_page.add_content(portlet)
   @last_page.publish!
   assert @last_page.published?
@@ -45,9 +45,9 @@ end
 
 When /^a page with a portlet that raises both a 404 and 403 error exists$/ do
   @last_page = create(:public_page)
-  @raises_not_found = create(:portlet, :code => 'raise ActiveRecord::RecordNotFound', :template=>"I shouldn't be shown.'")
+  @raises_not_found = create(:portlet, :code => 'raise ActiveRecord::RecordNotFound', :template => "I shouldn't be shown.'")
   @last_page.add_content(@raises_not_found)
-  @raises_access_denied = create(:portlet, :code => 'raise Cms::Errors::AccessDenied', :template=>"I shouldn't be shown.'")
+  @raises_access_denied = create(:portlet, :code => 'raise Cms::Errors::AccessDenied', :template => "I shouldn't be shown.'")
   @last_page.add_content(@raises_access_denied)
   @last_page.publish!
   assert @last_page.published?
@@ -55,21 +55,32 @@ end
 
 When /^a page with a portlet that raises both a 403 and any other error exists$/ do
   @last_page = create(:public_page)
-  @raises_not_found = create(:portlet, :code => 'raise "A Generic Error"', :template=>"I shouldn't be shown.'")
+  @raises_not_found = create(:portlet, :code => 'raise "A Generic Error"', :template => "I shouldn't be shown.'")
   @last_page.add_content(@raises_not_found)
-  @raises_access_denied = create(:portlet, :code => 'raise Cms::Errors::AccessDenied', :template=>"I shouldn't be shown.'")
+  @raises_access_denied = create(:portlet, :code => 'raise Cms::Errors::AccessDenied', :template => "I shouldn't be shown.'")
   @last_page.add_content(@raises_access_denied)
   @last_page.publish!
   assert @last_page.published?
 end
 
+Given /^a portlet that throws an unexpected error exists$/ do
+  @page = create(:public_page)
+  @portlet_render = DynamicPortlet.create!(:name => "Test", :connect_to_page_id => @page.id, :connect_to_container => "main", :template => '<p id="hi">hello</p>')
+  @portlet_raise_generic = DynamicPortlet.create!(:name => "Test", :connect_to_page_id => @page.id, :connect_to_container => "main", :code => 'raise Exception')
+  @page.publish!
+end
 Given /^there is a portlet that uses a helper$/ do
   @page_path = "/with-helper"
   @portlet = create(:portlet_with_helper, page_path: @page_path)
 
 end
 When /^I view that page$/ do
-  visit @page_path
+  if (@page_path)
+    visit @page_path
+  else
+    visit most_recently_created_page.path
+  end
+
 end
 
 Then /^I should see the portlet helper rendered in the view$/ do
@@ -84,4 +95,11 @@ end
 When /^a guest views that page$/ do
   logout
   visit @page_path
+end
+
+Then /^the page should show content but not the error$/ do
+  refute page.has_content?('Exception'), "Exception should not appear on the page"
+  refute page.has_content?('Error'), "The word 'Error' should not appear on the page"
+  assert page.has_content?('hello'), "Should see other content"
+  should_see_a_page_titled(most_recently_created_page.title)
 end