+++ test/unit/lib/content_block_test.rb
@@ -2,7 +2,7 @@ require 'test_helper'
 
 class ContentBlockTest < ActiveSupport::TestCase
   def setup
-    @block = create(:html_block, :name => "Test")
+    @block = create(:html_block, :name => "Test", :publish_on_save => false)
   end
 
   def test_publishing
@@ -17,8 +17,8 @@ class ContentBlockTest < ActiveSupport::TestCase
 
     assert @block.published?
 
-    assert @block.update_attributes(:name => "Whatever")
-    
+    assert @block.update_attributes(:name => "Whatever", :publish_on_save => false)
+
 
     assert !@block.live?
 
@@ -38,7 +38,7 @@ class ContentBlockTest < ActiveSupport::TestCase
     assert_equal 1, @block.version, "Block should keep itself at version 1"
     assert_equal 1, @block.versions.size, "Should only have the one original version of the block"
     assert_equal 'Created', @block.draft.version_comment
-    assert result , "Update with same attributes should still return true" 
+    assert result, "Update with same attributes should still return true"
   end
 
   def test_custom_revision_comment
@@ -46,7 +46,7 @@ class ContentBlockTest < ActiveSupport::TestCase
     assert_equal "Something Changed", @block.draft.version_comment
   end
 
-  
+
 end
 
 class SoftPublishingTest < ActiveSupport::TestCase
@@ -57,7 +57,7 @@ class SoftPublishingTest < ActiveSupport::TestCase
 
   test "deleted? should return true for deleted records, false otherwise" do
     assert_equal false, @block.deleted?
-    @block.destroy  
+    @block.destroy
     assert_equal true, @block.deleted?
   end
 
@@ -140,7 +140,7 @@ class VersionedContentBlockTest < ActiveSupport::TestCase
   end
 
   test "Getting a block as of a particular version shouldn't be considered a 'new record'." do
-    @block.update_attributes(:name=>"Changed", :publish_on_save=>true)
+    @block.update_attributes(:name => "Changed", :publish_on_save => true)
     @block.reload
 
     @v1 = @block.as_of_version(1)
@@ -148,12 +148,12 @@ class VersionedContentBlockTest < ActiveSupport::TestCase
 
   end
 
-  test 'Calling publish_on_save should not be sufficent to publish the block'do
+  test 'Calling publish_on_save should not be sufficent to publish the block' do
     @block.publish_on_save = true
     @block.save
 
     found = Cms::HtmlBlock.find(@block)
-    assert_equal 1, found.version 
+    assert_equal 1, found.version
   end
 
   test "Setting the 'publish' flag on a block, along with any other change, and saving it should mark that block as published." do
@@ -167,7 +167,7 @@ class VersionedContentBlockTest < ActiveSupport::TestCase
 
   def test_edit
     old_name = "Versioned Content Block"
-    new_name  = "New version of content block"
+    new_name = "New version of content block"
     @block.publish!
     @block.reload
     assert_equal @block.draft.name, old_name
@@ -183,7 +183,7 @@ class VersionedContentBlockTest < ActiveSupport::TestCase
 
   def test_revert
     old_name = "Versioned Content Block"
-    new_name  = "New version of content block"
+    new_name = "New version of content block"
     @block.publish!
     @block.reload
     assert_equal @block.draft.name, old_name
@@ -201,28 +201,41 @@ end
 
 class VersionedContentBlockConnectedToAPageTest < ActiveSupport::TestCase
   def setup
-    @page = create(:page, :section => root_section)
+    @page = create(:page, :section => root_section, :publish_on_save => false)
     @block = create(:html_block, :name => "Versioned Content Block")
     @page.create_connector(@block, "main")
     reset(:page, :block)
+
+
+  end
+
+  test "Adding a block to page" do
+    refute @page.published?, "The page should not be published yet."
+    assert_equal 2, @page.versions.size, "Should be two versions of the page"
+    pages = Cms::Page.connected_to(:connectable => @block, :version => @block.version).all
+    assert_equal [@page], pages, "The block should be connected to page"
+  end
+
+
+  test "updating block should not skip_callbacks" do
+    assert @block.update_attributes(:name => "something different", :publish_on_save => false)
+    assert_equal false, @block.skip_callbacks
+  end
+
+  test "updating block should put page into draft mode" do
+    @block.update_attributes(:name => "something different", :publish_on_save => false)
+    reset(:page)
+    refute @page.published?, "Updating a block should put any connected pages into draft mode"
   end
 
   def test_editing_connected_to_an_unpublished_page
     page_version_count = Cms::Page::Version.count
-    assert_equal 2, @page.versions.size, "Should be two versions of the page"
-    assert !@page.published?, "The page should not be published yet."
 
-    pages = Cms::Page.connected_to(:connectable => @block, :version => @block.version).all
-    assert_equal [@page], pages, "The block should be connected to page"
-
-
-    assert @block.update_attributes(:name => "something different")
-    assert_equal false, @block.skip_callbacks
+    assert @block.update_attributes(:name => "something different", :publish_on_save => false)
     assert_equal 2, @block.versions.size, "should be two versions of this block"
     reset(:page)
 
 
-    assert !@page.published?
     assert_equal 3, @page.versions.size, "Should be three versions of the page."
     assert_equal 3, @page.draft.version, "Draft version of page should be updated to v3 since its connected to the updated block."
     assert_incremented page_version_count, Cms::Page::Version.count
@@ -266,7 +279,7 @@ end
 
 class NonVersionedContentBlockConnectedToAPageTest < ActiveSupport::TestCase
   def setup
-    @page = create(:page, :section => root_section)
+    @page = create(:page, :section => root_section, :publish_on_save => false)
     @block = create(:non_versioned_block, :name => "Non-Versioned Non-Publishable Content Block")
     @page.create_connector(@block, "main")
     reset(:page, :block)